Create irecept launch class
